core/requirements_dev.txt
# Chamando o requirements do projeto.
-r requirements.txt

alabaster==0.7.12
appdirs==1.4.3
arrow==0.14.6
autopep8==1.4.4
Babel==2.7.0
dj-database-url==0.5.0
django-heroku==0.3.1
django-js-asset==1.2.2
docutils==0.15.2
imagesize==1.1.0
isort==4.3.21
packaging==16.8
pycparser==2.19
Pygments==2.4.2
pyparsing==2.2.0
PyYAML==5.1.2
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxjp.themes.dotted==0.1.2
sqlparse==0.3.0
xmltodict==0.12.0
beautifulsoup4==4.9.1
html5lib
